requirements.txt
Flask
SQLAlchemy
psycopg2
Flask-Migrate
webdriver-manager
Pillow
Stripe
gunicorn
openai
PyJWT
python-dotenv
Flask-Session
Jinja2
requests
Werkzeug
psycopg2-binary